Update create_multiple_sample_sheets.py
Deleted the sample sheet change for DRAGEN with use for the WGS sample sheet.  Since the BCL conversion with DRAGEN is not being implemented, the sample sheet for WGS has the sample FASTQ names and sample folder names reversed.  Right now, this means that the WGS portion of a run has to be redemultiplexed to make sure the sample fastq names and sample folder names are correct.  Hopefully, we can change the code back in the future to allow to call DRAGEN for WGS portion of the runs.